Moved user commands to the proper directory
