+++ TaskManager/src/Application/Command/User/AddUserCommand.php
@@ -12,7 +12,7 @@
 
 declare(strict_types=1);
 
-namespace Kreta\TaskManager\Application\User;
+namespace Kreta\TaskManager\Application\Command\User;
 
 class AddUserCommand
 {
